+++ Backtrace/Model/BacktraceReport.cs
@@ -9,7 +9,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 #if !NET35
-using System.Runtime.CompilerServices;
 using System.Runtime.ExceptionServices;
 #endif
 
@@ -71,6 +70,12 @@ namespace Backtrace.Model
         public List<string> AttachmentPaths { get; set; }
 
         /// <summary>
+        /// Current report exception stack
+        /// </summary>
+        [JsonProperty(PropertyName = "diagnosticStack")]
+        public List<BacktraceStackFrame> DiagnosticStack { get; set; }
+
+        /// <summary>
         /// Get or set minidump attachment path
         /// </summary>
         [JsonProperty(PropertyName = "minidumpFile")]
@@ -79,13 +84,9 @@ namespace Backtrace.Model
         /// <summary>
         /// Get an assembly where report was created (or should be created)
         /// </summary>
-        internal Assembly CallingAssembly { get; private set; }
+        internal Assembly CallingAssembly { get; set; }
 
-        /// <summary>
-        /// Current report exception stack
-        /// </summary>
-        [JsonProperty(PropertyName = "diagnosticStack")]
-        public List<DiagnosticStack> DiagnosticStack { get; private set; } = new List<DiagnosticStack>();
+        internal readonly bool _reflectionMethodName;
 
         /// <summary>
         /// Create new instance of Backtrace report to sending a report with custom client message
@@ -97,12 +98,11 @@ namespace Backtrace.Model
         public BacktraceReport(
             string message,
             Dictionary<string, object> attributes = null,
-            List<string> attachmentPaths = null)
+            List<string> attachmentPaths = null,
+            bool reflectionMethodName = true)
+            : this(null as Exception, attributes, attachmentPaths, reflectionMethodName)
         {
             Message = message;
-            Attributes = attributes ?? new Dictionary<string, object>();
-            AttachmentPaths = attachmentPaths ?? new List<string>();
-            SetCallingAppInformation();
         }
 
         /// <summary>
@@ -115,16 +115,16 @@ namespace Backtrace.Model
         public BacktraceReport(
             Exception exception,
             Dictionary<string, object> attributes = null,
-            List<string> attachmentPaths = null)
+            List<string> attachmentPaths = null,
+            bool reflectionMethodName = true)
         {
             Attributes = attributes ?? new Dictionary<string, object>();
             AttachmentPaths = attachmentPaths ?? new List<string>();
             Exception = exception;
             ExceptionTypeReport = exception != null;
             Classifier = ExceptionTypeReport ? exception.GetType().Name : string.Empty;
-            CallingAssembly = exception.GetExceptionSourceAssembly();
-            SetCallingAppInformation();
-
+            _reflectionMethodName = reflectionMethodName;
+            SetCallingAssemblyInformation();
         }
 
         /// <summary>
@@ -163,97 +163,12 @@ namespace Backtrace.Model
             return reportAttributes.Merge(attributes);
         }
 
-        /// <summary>
-        /// Set Calling Assembly and current thread stack trace property. 
-        /// CallingAssembly and StackTrace are necessary to prepare diagnostic JSON in BacktraceData class
-        /// </summary>
-        internal void SetCallingAppInformation()
+        internal void SetCallingAssemblyInformation()
         {
-            // generate stacktrace with file info
-            // if assembly have pbd files, diagnostic JSON will contain information about 
-            // line number and column number
-            var stackTrace = new StackTrace(true);
-            var stackFrames = stackTrace.GetFrames();
-            SetStacktraceInformation(stackFrames, true);
-
-            if (Exception == null)
-            {
-                return;
-            }
-            // add stack trace from exception
-            var head = DiagnosticStack.Any() ? DiagnosticStack[0] : null;
-            var generatedStack = Exception.GetExceptionStackFrames(head);
-            SetStacktraceInformation(generatedStack, false);
-
-            //Library didn't found Calling assembly
-            //The reason for this behaviour is because we throw exception from TaskScheduler
-            //or other method that don't generate valid stack trace
-            if (CallingAssembly == null)
-            {
-                CallingAssembly = Assembly.GetExecutingAssembly();
-            }
+            var stacktrace = new BacktraceStackTrace(Exception, _reflectionMethodName);
+            DiagnosticStack = stacktrace.StackFrames;
+            CallingAssembly = stacktrace.CallingAssembly;
         }
-
-        private void SetStacktraceInformation(StackFrame[] stackFrames, bool includeCallingAssembly, int startingIndex = 0)
-        {
-            // check if stack frames exists
-            if (stackFrames == null)
-            {
-                return;
-            }
-            var executedAssemblyName = Assembly.GetExecutingAssembly().FullName;
-            //if callingAssemblyFound is true, we dont need to found calling assembly in current stacktrace
-            bool callingAssemblyFound = !includeCallingAssembly;
-
-            foreach (var stackFrame in stackFrames)
-            {
-                var method = stackFrame.GetMethod();
-                var declaringType = method?.DeclaringType;
-                if (declaringType == null)
-                {
-                    //received invalid or unvailable stackframe
-                    continue;
-                }
-                Assembly assembly = declaringType.Assembly;
-                if (assembly == null)
-                {
-                    continue;
-                }
-                var assemblyName = assembly.FullName;
-                if (executedAssemblyName.Equals(assemblyName))
-                {
-                    // remove all system and microsoft stack frames 
-                    //if we add any stackframe to list this is mistake because we receive 
-                    //system or microsoft dll (for example async invoke)
-                    DiagnosticStack.Clear();
-                    startingIndex = 0;
-                    continue;
-                }
-                if (!callingAssemblyFound && ((!(SystemHelper.SystemAssembly(assembly)))
-                    || (CallingAssembly != null && assembly?.FullName == CallingAssembly.FullName)))
-                {
-                    callingAssemblyFound = true;
-                    CallingAssembly = assembly;
-                }
-#if !NET35
-                //test if current stack frame is generated by async state machine
-                var declaringTypeInfo = declaringType.GetTypeInfo();
-                var stateMachineFrame = SystemHelper.StateMachineFrame(declaringTypeInfo);
-                if (stateMachineFrame)
-                {
-                    continue;
-                }
-#endif
-                if (!callingAssemblyFound)
-                {
-                    continue;
-                }
-                var diagnosticStack = Model.JsonData.DiagnosticStack.Convert(stackFrame, assembly.GetName().Name, true);
-                DiagnosticStack.Insert(startingIndex, diagnosticStack);
-                startingIndex++;
-            }
-        }
-
         /// <summary>
         /// create a copy of BacktraceReport for inner exception object inside exception
         /// </summary>
@@ -268,7 +183,7 @@ namespace Backtrace.Model
             }
             var copy = (BacktraceReport)this.MemberwiseClone();
             copy.Exception = this.Exception.InnerException;
-            copy.SetCallingAppInformation();
+            copy.SetCallingAssemblyInformation();
             copy.Classifier = copy.Exception.GetType().Name;
             return copy;
         }