+++ Backtrace/Model/JsonData/ThreadData.cs
@@ -29,7 +29,7 @@ namespace Backtrace.Model.JsonData
         /// <summary>
         /// Create instance of ThreadData class to collect information about used threads
         /// </summary>
-        internal ThreadData(Assembly callingAssembly, IEnumerable<DiagnosticStack> exceptionStack)
+        internal ThreadData(Assembly callingAssembly, IEnumerable<BacktraceStackFrame> exceptionStack)
         {
 #if NET45
             //use available in .NET 4.5 api to find stack trace of all available managed threads
@@ -46,13 +46,13 @@ namespace Backtrace.Model.JsonData
         /// Generate information for current thread
         /// </summary>
         /// <param name="exceptionStack">Current BacktraceReport exception stack</param>
-        private void GenerateCurrentThreadInformation(IEnumerable<DiagnosticStack> exceptionStack)
+        private void GenerateCurrentThreadInformation(IEnumerable<BacktraceStackFrame> exceptionStack)
         {
             var current = Thread.CurrentThread;
             //get current thread id
             string generatedMainThreadId = current.GenerateValidThreadName().ToLower();
 
-            ThreadInformations[generatedMainThreadId] = new ThreadInformation(current, exceptionStack);
+            ThreadInformations[generatedMainThreadId] = new ThreadInformation(current, exceptionStack, true);
             //set currentThreadId
             MainThread = generatedMainThreadId;
         }
@@ -126,7 +126,15 @@ namespace Backtrace.Model.JsonData
                     {
                         threadName = thread.ManagedThreadId.ToString();
                     }
-                    var frames = DiagnosticStack.Convert(thread.StackTrace);
+                    var frames = new List<BacktraceStackFrame>();
+                    foreach (var frame in thread.StackTrace)
+                    {
+                        if(frame.Method == null)
+                        {
+                            continue;
+                        }
+                        frames.Add(new BacktraceStackFrame(frame));
+                    }
                     ThreadInformations[threadName] = new ThreadInformation(threadName, false, frames);
                 }
             }